+++ toys/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Release History
 
+### 0.3.6 / TBD
+
 ### 0.3.5 / 2018-05-15
 
 * CHANGED: Flag and arg blocks in the DSL have an interface more similar to the rest of the DSL.